backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosmtplib==1.1.7
alembic==1.9.0
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==3.4.8
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.87.0
fastapi-cache2==0.2.1
fastapi-jwt-auth==0.5.0
fastapi-mail==1.2.2
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
multidict==6.0.4
orjson==3.8.3
passlib==1.7.4
pendulum==2.1.2
psycopg2==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytzdata==2020.1
PyYAML==6.0
redis==4.5.5
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.21.0
typing_extensions==4.4.0
ujson==5.6.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
yarl==1.9.2
